yang/requirements.txt
pyang<=2.3.2
pyangbind<=0.8.1
PyYAML<=5.3.1
jinja2<=2.11.2